add execution time in miliseconds on api exec add option to obtain a session stream as a utf8 content instead of event stream list refactor sessions route
