+++ .env.sample
@@ -40,7 +40,7 @@ ADMIN_USERNAME=admin
 # PYROSCOPE_AUTH_TOKEN=noop
 
 # xtdb connection configuration
-PG_HOST=
+PG_HOST=host.docker.internal
 PG_USER=hoopdevuser
 PG_PASSWORD=1a2b3c4d
 PG_DB=hoopdevdb
